update radius type to float , increase preision of pi

+++ circle.c
@@ -2,10 +2,10 @@
 #define PI 3.14
 int main()
 {
-    int radius;
+    float radius;
     float area,circumference;
     printf("Enter the radius of the circle:");
-    scanf("%d",&radius);
+    scanf("%f",&radius);
     area = PI *radius*radius;
     printf("Area of circle: %0.3f \n",area);
     circumference= 2*PI*radius;
